django-rest/requirements.txt
# Requirements for the Django rest framework project
Django>=5.2.3
mysqlclient
uvicorn
gunicorn
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
pymysql
fastapi
asyncio
aiomysql
django-filter
Pillow